# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"KataLevel": {
				"enum": [
					"Basic",
					"Medium",
					"High"
				],
				"type": "string"
			},
			"IKata": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/KataLevel"
					},
					"intents": {
						"type": "number",
						"format": "double"
					},
					"stars": {
						"type": "number",
						"format": "double"
					},
					"creator": {
						"type": "string"
					},
					"participants": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"solution": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"level",
					"intents",
					"stars",
					"creator",
					"participants",
					"solution"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DateResponse": {
				"properties": {
					"Date": {},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"Date",
					"message"
				],
				"type": "object"
			},
			"BasicResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"description": "Para definir la estructura de mis datos: devuelve un json"
			},
			"IUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"age",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuth": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend-code-verifier",
		"version": "1.0.0",
		"description": "Node - Express - Typescript - Swagger ",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "LoreLor"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/katas": {
			"get": {
				"operationId": "GetKataByLevel",
				"responses": {
					"200": {
						"description": "detail of kata",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to get Kata By Level in the Collection \"katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query param from Id of kata",
						"in": "query",
						"name": "level",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/KataLevel"
						}
					}
				]
			},
			"post": {
				"operationId": "KataCreate",
				"responses": {
					"200": {
						"description": "the new kata created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to post a new Kata in the Collection \"katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "object with data from body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IKata",
								"description": "object with data from body"
							}
						}
					}
				}
			}
		},
		"/api/katas/{id}": {
			"get": {
				"operationId": "GetKataById",
				"responses": {
					"200": {
						"description": "detail of kata",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to get Kata By Id in the Collection \"katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"description": "query param from Id of kata",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "KataUpdate",
				"responses": {
					"200": {
						"description": "Kata updating in DB",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to update an Kata in the Collection \"katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "recives user Id and update Object user",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IKata",
								"description": "recives user Id and update Object user"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "KataDelete",
				"responses": {
					"200": {
						"description": "delete kata by id  message",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to delete Kata in the Collection \"katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id of kata to delete (optional)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "All Users or User by Id",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to retrieve the Users in the Collection \"users\" of DB",
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Id of user to retriev (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "delete user by id  message",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to delete User in the Collection \"users\" of DB",
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id of user to delete (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "User updating in DB",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to update an User in the Collection \"users\" of DB",
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "recives user Id and update Object user",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "recives user Id and update Object user"
							}
						}
					}
				}
			}
		},
		"/api/goodbye": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"GoodbyeController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/hello": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Promise de BasicResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponse"
								}
							}
						}
					}
				},
				"description": "Endpoint que devuelve un mensaje Hello {name} en json",
				"tags": [
					"HelloController"
				],
				"security": [],
				"parameters": [
					{
						"description": "nombre dado desde query",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/auth/register": {
			"post": {
				"operationId": "UserRegister",
				"responses": {
					"200": {
						"description": "User created in DB",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to post a new User Authenticated in the Collection \"users\" of DB",
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Object created",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser",
								"description": "Object created"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "UserLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuth"
							}
						}
					}
				}
			}
		},
		"/api/auth/userData": {
			"get": {
				"operationId": "UserData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}